Practice a common 'modal' style
The blue icon is on the left. There is equal space on either side of the
icon. There is padding around the edge of the modal. The header, text,
and buttons are aligned with each other. The header is bold and slightly
larger than the surrounding text. The close button is vertically
aligned with the header and is in the top-right.