Style an entire page
The header is at the top of the page and the footer is at the bottom.
They both have padding. Header and footer links are pushed to either
side. Gap is used to create space between them. The footer has a light
gray background. Links are styled. The main contents are centered in the
screen.